refactor(navbar): extract navbar logic to custom hook
