+++ src/components/users/home/PosterCard.jsx
@@ -41,7 +41,7 @@ export default function PosterCard() {
               <button
                 type="button"
                 className="btn bg-primary-4 text-neutral-5 text-xs mt-4 sm:mt-8"
-                onClick={() => navigate("/search")}
+                onClick={() => navigate("/products")}
               >
                 Beli Sekarang
               </button>
